+++ database.js
@@ -38,9 +38,10 @@ const pool = mysql.createPool({
     user: process.env.MYSQL_USER,
     password: process.env.MYSQL_PASSWORD,
     database: process.env.MYSQL_DATABASE
-});
+}).promise()
 
 pool.getConnection((err,connection)=> {
+    console.log('Function 1 called');
     if(err) {
         console.log("AAAAHHHHH");
         databaseExists = false;
@@ -56,9 +57,9 @@ pool.getConnection((err,connection)=> {
  * get a list of all users
  * @returns list of all users, each element is a User object with fields username, pass, and id
  */
-export function getUsers() {
+export async function getUsers() {
     if (!databaseExists) return undefined
-    const result = pool.query("SELECT * FROM users");
+    const result = await pool.query("SELECT * FROM users");
     const rows = result[0];
     return rows;
 }
@@ -68,9 +69,9 @@ export function getUsers() {
  * @param {number} userId 
  * @returns User object with fields username, pass, and id
  */
-export function getUser(userId) {
+export async function getUser(userId) {
     if (!databaseExists) return undefined
-    const [rows] = pool.query(`
+    const [rows] = await pool.query(`
     SELECT * 
     FROM users
     WHERE id = ?
@@ -83,7 +84,7 @@ export function getUser(userId) {
  * @param {string} username 
  * @returns User object, with fields username, pass, and id
  */
-export function getUserWithUsername(username) {
+export async function getUserWithUsername(username) {
     if (!databaseExists) return undefined
     const [rows] = pool.query(`
     SELECT *
@@ -99,7 +100,7 @@ export function getUserWithUsername(username) {
  * @param {string} password 
  * @returns the id of the User if it exists, or undefined if it does not exist
  */
-export function userExists(username, password) {
+export async function userExists(username, password) {
     if (!databaseExists) return undefined
     const [rows] = pool.query(`
     SELECT *
@@ -116,7 +117,7 @@ export function userExists(username, password) {
  * @param {string} pass 
  * @returns struct with fields id, username, and pass
  */
-export function createUser(username, pass) {
+export async function createUser(username, pass) {
     if (!databaseExists) return undefined
     const result = pool.query(`
     INSERT INTO users (username, pass)
@@ -134,7 +135,7 @@ export function createUser(username, pass) {
  * @param {number} userId 
  * @returns struct with fields id, username, pass of the user which got deleted. if it doesn't exists, returns undefined.
  */
-export function removeUser(userId) {
+export async function removeUser(userId) {
     if (!databaseExists) return undefined
     const item = pool.query(`
     SELECT *
@@ -159,7 +160,7 @@ export function removeUser(userId) {
  * @param {number} budgetId
  * @returns budget row, or undefined if it does not exist
  * */ 
-export function getBudget(budgetId) {
+export async function getBudget(budgetId) {
     if (!databaseExists) return undefined
     const [rows] = pool.query(`
     SELECT *
@@ -176,9 +177,9 @@ export function getBudget(budgetId) {
  * @param {number} year 
  * @returns budget row with the specified quarter and year, or undefined if it does not exist
  */
-export function getBudgetForUserId(userId, quarter, year) {
+export async function getBudgetForUserId(userId, quarter, year) {
     if (!databaseExists) return undefined
-    const [rows] = pool.query(`
+    const [rows] = await pool.query(`
     SELECT * 
     FROM budgets
     WHERE userId=? AND quarter=? AND year=?
@@ -191,9 +192,9 @@ export function getBudgetForUserId(userId, quarter, year) {
  * @param {number} userId 
  * @returns budget row(s) that the user userId owns
  */
-export function getAllBudgetsForUserId(userId) {
+export async function getAllBudgetsForUserId(userId) {
     if (!databaseExists) return undefined
-    const [rows] = pool.query(`
+    const [rows] = await pool.query(`
     SELECT *
     FROM budgets
     WHERE userId=?
@@ -208,9 +209,9 @@ export function getAllBudgetsForUserId(userId) {
  * @param {number} year 
  * @returns array of budget id, quarter, year
  */
-export function createBudget(userId, quarter, year)  {
+export async function createBudget(userId, quarter, year)  {
     if (!databaseExists) return undefined
-    const result = pool.query(`
+    const result = await pool.query(`
     INSERT INTO
     budgets (userId, quarter, year)
     VALUES (?, ?, ?)
@@ -229,9 +230,9 @@ export function createBudget(userId, quarter, year)  {
  * @param {number} budgetId 
  * @returns list of Expenses. each element is a struct with fields id, amount, and type
  */
-export function getExpenses(budgetId) {
+export async function getExpenses(budgetId) {
     if (!databaseExists) return undefined
-    const expenses = pool.query(`
+    const expenses = await pool.query(`
     SELECT * 
     FROM expenses
     WHERE budgetId = ?
@@ -246,9 +247,9 @@ export function getExpenses(budgetId) {
  * @param {ExpenseType} type 
  * @returns the Expense that was created (struct contiaining id, amount, and type)
  */
-export function createExpense(budgetId, amount, type) {
+export async function createExpense(budgetId, amount, type) {
     if (!databaseExists) return undefined
-    const expense = pool.query(`
+    const expense = await pool.query(`
     INSERT
     INTO expenses (budgetId, amount, type)
     VALUES (?, ?, ?)
@@ -265,9 +266,9 @@ export function createExpense(budgetId, amount, type) {
  * @param {number} budgetId id of the budget
  * @returns a list of all Incomes for that budget. each Income is a struct with fields id, amount, and type
  */
-export function getIncomes(budgetId) {
+export async function getIncomes(budgetId) {
     if (!databaseExists) return undefined
-    const incomes = pool.query(`
+    const incomes = await pool.query(`
     SELECT *
     FROM incomes
     WHERE budgetId = ?
@@ -282,9 +283,9 @@ export function getIncomes(budgetId) {
  * @param {IncomeType} type 0: 'unassigned', 1: 'grant', 2: 'loan', 3: 'wages', 4: 'family'
  * @returns struct with id, amount, and type of the Income that was created
  */
-export function createIncome(budgetId, amount, type) {
+export async function createIncome(budgetId, amount, type) {
     if (!databaseExists) return undefined
-    const income = pool.query(`
+    const income = await pool.query(`
     INSERT
     INTO incomes (budgetId, amount, type)
     VALUES (?, ?, ?)
@@ -302,7 +303,7 @@ export function createIncome(budgetId, amount, type) {
  * @param {number} amount savings amount
  * @returns Savings object that contains the id of the savings that was inserted, and the amount
  */
-export function createSavings(budgetId, amount) {
+export async function createSavings(budgetId, amount) {
     if (!databaseExists) return undefined
     const savings = pool.query(`
     INSERT
@@ -320,7 +321,7 @@ export function createSavings(budgetId, amount) {
  * @param {number} budgetId id of the budget
  * @param {number} amount numerical amount to update to
  */
-export function updateSavings(budgetId, amount) {
+export async function updateSavings(budgetId, amount) {
     if (!databaseExists) return undefined
     const result = pool.query(`
     SELECT
@@ -332,3 +333,5 @@ export function updateSavings(budgetId, amount) {
 // savings goal
 // 
 
+// const user = await getUsers();
+// console.log(user);
