+++ static/budget-report.html
@@ -10,11 +10,9 @@
     <h1 id="title">
         Budget Report
     </h1>
- 
-    <div class="stats" id="amount_used_label">Amount Used: $<span id="amount_used_value">0</span></div>
-    <div class="stats" id="remaining_fund_label">Remaining Fund: $<span id="remaining_fund_value">0</span></div>
-    <div class="stats" id="out_of_pocket_label">Out of Pocket: $<span id="out_of_pocket_value">0</span></div>
-    
+
+     <div id="budgetDataContainer"></div>
+
     
     <div class="graphs-container">
         <div id="bargraph" class="chart-container"></div>
